Try to fix the deploy script

+++ utils/update-gh-pages.sh
@@ -22,25 +22,27 @@ if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then
   git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/lupino3/edumips64.git gh-pages > /dev/null
   echo "done."
   
-  GIT_BRANCH=$(git rev-parse --abbrev-ref HEAD)
+  echo "Git branch: $TRAVIS_BRANCH"
 
-  if [ "$GIT_BRANCH" == "master" ]; then 
+  if [ "$TRAVIS_BRANCH" == "master" ]; then 
     echo -n "Building JAR.. "
     ant latest-jar > /dev/null
     cp edumips64-latest.jar $HOME > /dev/null
     echo "done."
-	cd gh-pages
-	mv $HOME/edumips64-latest.jar .
+	  cd gh-pages
+	  mv $HOME/edumips64-latest.jar .
+  elif ["$TRAVIS_BRANCH" == "webui-prototype"]; then
+    echo -n "Copying web UI..."
+	  cd gh-pages
+    rm -rf webui  
+    mkdir webui
+    cp -Rf $HOME/webui/* ./webui  
+    echo "done."
   else
-	cd gh-pages
     echo "Non-master branch: $GIT_BRANCH. Not updating the JAR."
+    return 0
   fi
 
-  echo -n "Copying web UI..."
-  rm -rf webui  
-  mkdir webui
-  cp -Rf $HOME/webui/* ./webui  
-  
   # Add, commit and push files.
   echo -n "Committing files.. "
   git add -f .
